+++ src/UserPermissionsTrait.php
@@ -25,14 +25,7 @@ trait UserPermissionsTrait
             UserPermissions::EVENT_REGISTER_PERMISSIONS,
             function(RegisterUserPermissionsEvent $event) {
 
-                $event->permissions[ $this->name ] = [
-                    Reports::PERMISSION_ACCESS => ['label' => Craft::t('reports', 'Access Reports')],
-                    Reports::PERMISSION_CREATE => ['label' => Craft::t('reports', 'Create Reports')],
-                    Reports::PERMISSION_EDIT   => ['label' => Craft::t('reports', 'Edit Reports')],
-                    Reports::PERMISSION_RUN    => ['label' => Craft::t('reports', 'Run Reports')],
-                    Reports::PERMISSION_EXPORT => ['label' => Craft::t('reports', 'Export Reports')],
-                    Reports::PERMISSION_DELETE => ['label' => Craft::t('reports', 'Delete Reports')],
-                ];
+                $event->permissions[ $this->name ] = $this->_getPermissions();
             }
         );
     }
@@ -42,16 +35,16 @@ trait UserPermissionsTrait
      */
     public function getCpNavItem()
     {
-        $navItem = parent::getCpNavItem();
+        $navItem          = parent::getCpNavItem();
         $navItem['label'] = $this->getPluginName();
-        $subNav = [];
+        $subNav           = [];
 
         /** @var User $currentUser */
         $currentUser = Craft::$app->getUser()->getIdentity();
 
         $permissions = [
             [
-                'type'   => Reports::PERMISSION_ACCESS,
+                'type'   => [Reports::PERMISSION_MANAGE_REPORTS, Reports::PERMISSION_RUN_REPORTS],
                 'id'     => 'reports',
                 'config' => [
                     'label' => 'Reports',
@@ -59,7 +52,7 @@ trait UserPermissionsTrait
                 ],
             ],
             [
-                'type'   => Reports::PERMISSION_TARGETS,
+                'type'   => Reports::PERMISSION_MANAGE_TARGETS,
                 'id'     => 'targets',
                 'config' => [
                     'label' => 'Report Targets',
@@ -68,8 +61,18 @@ trait UserPermissionsTrait
             ],
         ];
 
+        $checkPermissions = function($permissions) use ($currentUser) {
+            $results = array_map(function($permission) use ($currentUser) {
+                return $currentUser->can($permission);
+            }, $permissions);
+
+            return \in_array(true, $results);
+        };
+
         foreach ($permissions as $permission) {
-            if ($currentUser->can($permission['type'])) {
+            $canAccess = !\is_array($permission['type']) ? $currentUser->can($permission['type']) : $checkPermissions($permission['type']);
+
+            if ($canAccess) {
                 $subNav[ $permission['id'] ] = $permission['config'];
             }
         }
@@ -81,4 +84,20 @@ trait UserPermissionsTrait
         return $navItem;
     }
 
+    private function _getPermissions()
+    {
+        $permissions = [
+            Reports::PERMISSION_RUN_REPORTS    => 'Run Reports',
+            Reports::PERMISSION_MANAGE_REPORTS => 'Manage Reports',
+            Reports::PERMISSION_MANAGE_TARGETS => 'Manage Export Targets',
+        ];
+
+        $result = [];
+
+        foreach ($permissions as $key => $label) {
+            $result[ $key ] = ['label' => Craft::t('reports', $label)];
+        }
+
+        return $result;
+    }
 }